+++ src/components/ConfigUi/Checkbox/index.js
@@ -2,10 +2,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
 export default class Checkbox extends Component {
-  constructor() {
-    super();
+  constructor(props) {
+    super(props);
+    const status = props.indeterminate === 'true' ? 'indeterminate' : 'off';
     this.state = {
-      status: 'indeterminate'
+      status,
     };
   }
 
@@ -20,22 +21,34 @@ export default class Checkbox extends Component {
   }
 
   toggleState = () => {
-    if(this.state.status === 'indeterminate') {
-      this.setState({status: 'on'});
+    if (this.state.status === 'indeterminate') {
+      this.setState({ status: 'on' });
       return this.props.onChange('on');
-    } else if(this.state.status === 'on') {
-      this.setState({status: 'off'});
+    } else if (this.state.status === 'on') {
+      this.setState({ status: 'off' });
       return this.props.onChange('off');
-    } else if(this.state.status === 'off') {
-      this.setState({status: 'indeterminate'});
-      this.el.indeterminate = this.props.indeterminate;
-      return this.props.onChange('indeterminate');
+    } else if (this.state.status === 'off') {
+      if (this.props.indeterminate === 'true') {
+        this.setState({ status: 'indeterminate' });
+        this.el.indeterminate = this.props.indeterminate;
+        return this.props.onChange('indeterminate');
+      } else {
+        this.setState({ status: 'on' });
+        return this.props.onChange('on');
+      }
     }
-  }
+  };
 
   render() {
     return (
-      <input className="read-only" {...this.props} type="checkbox" ref={el => this.el = el} onChange={this.toggleState} checked={this.state.status === 'on'}  />
+      <input
+        className="read-only"
+        {...this.props}
+        type="checkbox"
+        ref={el => (this.el = el)}
+        onChange={this.toggleState}
+        checked={this.state.status === 'on'}
+      />
     );
   }
 }